refactor(web): migrate object-flatten to feature folder
