+++ apps/web/src/tools/object-flatten/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { ObjectFlatten } from "./ui";
+
+export const tool: ToolModule = { id: "object-flatten", Component: ObjectFlatten };
