+++ apps/web/src/tools/object-flatten/object-flatten.ts
